Improve pseudo tool tips
Handle text that is too wide for the window, the same as we handle
multi-line text: Show left-justified on following line(s).

Add a fudge factor for when we need to right-justify; w/o that we
sometimes still ran just off right edge.

Instead of point-at-{bol eol} use {beginning end}-of-visual-line,
which gives good results when visual-line-mode is enabled. (When it is
not enabled, this is probably a bit slower, but AFAICT it still
works.)

Note: These changes arose from me reviewing the code for issue #514,
and this commit is initially on branch named after that issue. But,
AFAICT this does nothing to address whatever is the underlying cause
of that seemingly infrequent problem.